Get rid of old graphics demo
