--- tests/graphics_test/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_executable(graphics_test
-    main.cpp
-)
-
-target_link_libraries(graphics_test PRIVATE phys_graphics)
